Rename todo to shopping list
