+++ AwesomeProject/src/screens/componets/Explore/Todo.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import {StyleSheet, Text, View} from 'react-native';
-import AddTodo from '../Todo/AddTodo';
+import AddTodo from '../Todo/AddList';
 import VisibleTodos from '../Todo/VisibleTodos';
 
 class Todo extends Component {
